fix: TaskCmd.Restart if the process has already started #40
It seem to happen because TaskCmd.Restart executes TaskCmd.Start before waiting all the scoped processes are finished, which may remain process's state Stopping.

In this case, TaskCmd.Start will skip to start the process because it's not start-able and checkStatus in TaskCmd.Start will fail.

solution: create the StopAndWait function to wait all process are stopped.